Объединяет viewmodels
