+++ featureDice/src/main/java/com/bestscore/featuredice/DiceMode.kt
@@ -1,4 +1,4 @@
-package com.bestscore.featuredice.dice
+package com.bestscore.featuredice
 
 enum class DiceMode {
     MODE_1D4,
