+++ featureDice/src/main/java/com/bestscore/featuredice/TossResult.kt
@@ -1,4 +1,4 @@
-package com.bestscore.featuredice.coin
+package com.bestscore.featuredice
 
 enum class TossResult {
     HEADS,
